requirements.txt
matplotlib==3.5.2
pandas==1.4.3
scipy==1.8.1
mne==1.0.3
brainflow==5.0.1
google-cloud-storage==2.4.0
Flask==2.1.0
gunicorn==20.1.0
Jinja2==3.0.3